Move code of website into components
