+++ website/src/components/HomepageHeader/index.js
@@ -1,0 +1,43 @@
+import React from 'react';
+import clsx from 'clsx';
+import Link from '@docusaurus/Link';
+import useDocusaurusContext from '@docusaurus/useDocusaurusContext';
+import logoDataUri from '@site/static/img/logo.png';
+
+import styles from './styles.module.css';
+
+export default function HomepageHeader() {
+  const { siteConfig } = useDocusaurusContext();
+  return (
+    <header className={clsx('hero hero--primary', styles.heroBanner)}>
+      <div className="container">
+        <h1 className={clsx('hero__title', styles.mainTitle)}>
+          <img src={logoDataUri} alt={siteConfig.title} className={styles.mainTitleLogo} width="570px" />
+        </h1>
+        <p className={clsx('hero__subtitle', styles.subTitle)}>
+          {siteConfig.tagline}
+          <div className={styles.taglineBadges}>
+            <a href="https://www.npmjs.com/package/fast-check" target="_blank" rel="noopener">
+              <img src="https://badge.fury.io/js/fast-check.svg" alt="npm version" height={20} loading="lazy" />
+            </a>
+            <a href="https://www.npmjs.com/package/fast-check" target="_blank" rel="noopener">
+              <img src="https://img.shields.io/npm/dm/fast-check" alt="monthly downloads" height={20} loading="lazy" />
+            </a>
+            <a href="https://github.com/dubzzz/fast-check/stargazers" target="_blank" rel="noopener">
+              <img
+                src="https://img.shields.io/github/stars/dubzzz/fast-check?style=social"
+                height={20}
+                loading="lazy"
+              />
+            </a>
+          </div>
+        </p>
+        <div className={styles.quickNavigationButtons}>
+          <Link className="button button--secondary button--lg" to="/docs/tutorials/quick-start/">
+            Quick Start - 5min ⏱️
+          </Link>
+        </div>
+      </div>
+    </header>
+  );
+}
