+++ website/src/components/HomepageSponsors/styles.module.css
@@ -1,0 +1,3 @@
+.allSponsorsWrapper {
+  align-self: center;
+}
